+++ app/src/main/java/vadimerenkov/aucards/screens/CardFullscreen.kt
@@ -2,7 +2,6 @@
 
 package vadimerenkov.aucards.screens
 
-//import vadimerenkov.aucards.CreateStarters
 import android.content.pm.ActivityInfo
 import android.provider.Settings
 import android.util.Log
@@ -13,54 +12,42 @@ import androidx.compose.animation.SharedTransitionScope
 import androidx.compose.animation.animateColorAsState
 import androidx.compose.animation.scaleIn
 import androidx.compose.animation.scaleOut
-import androidx.compose.foundation.background
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.ExperimentalLayoutApi
-import androidx.compose.foundation.layout.FlowRow
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.WindowInsets
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.isImeVisible
 import androidx.compose.foundation.layout.navigationBarsPadding
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.text.KeyboardOptions
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.Close
 import androidx.compose.material.icons.filled.Done
-import androidx.compose.material3.DropdownMenu
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.Icon
 import androidx.compose.material3.IconButton
 import androidx.compose.material3.IconButtonDefaults
-import androidx.compose.material3.LocalTextStyle
 import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.OutlinedTextField
 import androidx.compose.material3.Surface
-import androidx.compose.material3.Tab
-import androidx.compose.material3.TabRow
 import androidx.compose.material3.Text
 import androidx.compose.material3.TextField
 import androidx.compose.material3.TextFieldDefaults
-import androidx.compose.material3.minimumInteractiveComponentSize
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.DisposableEffect
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableIntStateOf
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
 import androidx.compose.ui.focus.FocusRequester
 import androidx.compose.ui.focus.focusRequester
 import androidx.compose.ui.graphics.Color
@@ -72,23 +59,20 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.input.ImeAction
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.DpOffset
+import androidx.compose.ui.unit.DpSize
 import androidx.compose.ui.unit.dp
-import androidx.compose.ui.window.PopupProperties
 import androidx.core.view.WindowCompat
 import androidx.core.view.WindowInsetsCompat
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import androidx.lifecycle.viewmodel.compose.viewModel
-import com.github.skydoves.colorpicker.compose.BrightnessSlider
-import com.github.skydoves.colorpicker.compose.HsvColorPicker
-import com.github.skydoves.colorpicker.compose.rememberColorPickerController
 import kotlinx.coroutines.Dispatchers
 import vadimerenkov.aucards.R
 import vadimerenkov.aucards.ViewModelFactory
 import vadimerenkov.aucards.data.Aucard
 import vadimerenkov.aucards.ui.CardState
 import vadimerenkov.aucards.ui.CardViewModel
+import vadimerenkov.aucards.ui.ColorPickerPopup
 import vadimerenkov.aucards.ui.ContentType
-import vadimerenkov.aucards.ui.Palette
 import vadimerenkov.aucards.ui.SharedContentStateKey
 import vadimerenkov.aucards.ui.calculateContentColor
 
@@ -233,7 +217,6 @@ private fun SharedTransitionScope.EditScreen(
 	textContentState: SharedTransitionScope.SharedContentState,
 	modifier: Modifier = Modifier
 ) {
-	val colorController = rememberColorPickerController()
 	var colorPaletteOpen by remember { mutableStateOf(false) }
 	val focusRequester = remember { FocusRequester() }
 
@@ -345,102 +328,19 @@ private fun SharedTransitionScope.EditScreen(
 						)
 					}
 
-					DropdownMenu(
-						offset = DpOffset((-50).dp, 0.dp),
+					ColorPickerPopup(
+						isOpen = colorPaletteOpen,
 						onDismissRequest = { colorPaletteOpen = false },
-						properties = PopupProperties(
-							//clippingEnabled = false
-						),
-						expanded = colorPaletteOpen
-					) {
-						var selectedTab by remember { mutableIntStateOf(0) }
-						TabRow(
-							selectedTabIndex = selectedTab,
-							modifier = Modifier
-								.width(300.dp)
-								.height(50.dp)
-						) {
-							Tab(
-								selected = selectedTab == 0,
-								onClick = { selectedTab = 0 }
-							) {
-								Text(stringResource(R.string.palette))
-							}
-							Tab(
-								selected = selectedTab == 1,
-								onClick = { selectedTab = 1 }
-							) {
-								Text(stringResource(R.string.custom_color))
-							}
-						}
+						selectedColor = state.aucard.color,
+						onColorSelected = onColorChange,
+						selectedHexCode = state.hexColor,
+						onHexCodeChanged = onHexChange,
+						isHexCodeValid = state.isHexCodeValid,
+						contentColor = contentColor,
+						offset = DpOffset(-(50).dp, 0.dp),
+						tabRowSize = DpSize(250.dp, 50.dp)
+					)
 
-						if (selectedTab == 0) {
-							FlowRow(
-								maxItemsInEachRow = 4,
-								modifier = Modifier
-									.padding(8.dp)
-									.align(Alignment.CenterHorizontally)
-							) {
-								Palette.colors.forEach { color ->
-									Box(
-										modifier = Modifier
-											.padding(4.dp)
-											.clip(MaterialTheme.shapes.medium)
-											.background(color)
-											.minimumInteractiveComponentSize()
-											.clickable(
-												onClick = { onColorChange(color) }
-											)
-									) {
-										if (color == state.aucard.color) {
-											Icon(
-												imageVector = Icons.Default.Done,
-												contentDescription = color.toString(),
-												tint = contentColor
-											)
-										}
-									}
-								}
-							}
-						}
-						if (selectedTab == 1) {
-							HsvColorPicker(
-								controller = colorController,
-								onColorChanged = {
-									onColorChange(it.color)
-									onHexChange(it.hexCode)
-								},
-								initialColor = state.aucard.color,
-								modifier = Modifier
-									.padding(8.dp)
-									.size(200.dp)
-									.align(Alignment.CenterHorizontally)
-							)
-							BrightnessSlider(
-								controller = colorController,
-								initialColor = state.aucard.color,
-								wheelRadius = 8.dp,
-								wheelColor = contentColor,
-								modifier = Modifier
-									.padding(horizontal = 16.dp)
-									.fillMaxWidth()
-									.height(24.dp)
-							)
-							OutlinedTextField(
-								value = state.hexColor.uppercase(),
-								isError = !state.isHexCodeValid,
-								singleLine = true,
-								textStyle = LocalTextStyle.current.copy(
-									textAlign = TextAlign.Center
-								),
-								onValueChange = onHexChange,
-								modifier = Modifier
-									.padding(8.dp)
-									.width(120.dp)
-									.align(Alignment.CenterHorizontally)
-							)
-						}
-					}
 				}
 			}
 		}
